refactor(ssa):reuse the code of sf export for risk export
